Refactor operator execution and environment handling
- Replaced `experimental_run_batch_env` with `SequentialBatchExecutor` for clearer and more modular batch execution.
- Moved `OperationRunner` to `operator.rs` and introduced `executor.rs` for batch executor logic.
- Simplified batch schema application by directly using `SequentialBatchExecutor`.
