+++ .run/Run firehose_loadtest.run.xml
@@ -1,7 +1,7 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="Run firehose_loadtest" type="CargoCommandRunConfiguration" factoryName="Cargo Command">
     <option name="buildProfileId" value="release" />
-    <option name="command" value="run --package firehose_loadtest --bin firehose_loadtest --release" />
+    <option name="command" value="run --package firehose_loadtest --bin firehose_loadtest" />
     <option name="workingDirectory" value="file://$PROJECT_DIR$" />
     <envs>
       <env name="CINIC10_PATH" value="/media/Data/CINIC-10" />
